Should I really be calling this "rubber ducking"?
The closest term I can think to fit it is "rubber ducking" since the essence of it is writing out and explaining my design ideas/options and their problems and working towards solutions that way; by coordinating my thoughts; but it's really just a design process that takes the form of a text document for convenience (the convenience being because it's highly impractical to store all this design information in my head, even though writing it out and reading it is indeed tedious and time consuming; I've found that the alternative is *much* more time consuming (and much more frustrating)). The form of a text document is also useful in conjunction with Git as I can see previous versions of it that way and so I don't have to worry about deleting potentially important things if they're saved in a previous commit.